refactor(switch): rename ignored spec files
