+++ libs/ui/switch/brain/jest.config.ts
@@ -13,7 +13,7 @@ export default {
 		],
 	},
 	transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
-	testPathIgnorePatterns: ['<rootDir>/src/lib/brn-switch-ng-model.component.spec.ts'],
+	testPathIgnorePatterns: ['<rootDir>/src/lib/brn-switch-ng-model.component.ignore.spec.ts'],
 	snapshotSerializers: [
 		'jest-preset-angular/build/serializers/no-ng-attributes',
 		'jest-preset-angular/build/serializers/ng-snapshot',
